Tests/LibWeb/Text/expected/URL/search-params-with-plus-in-value.txt

key 'key'
value 'value1 value2'
